feat: v1.1.7 more git config for initial .extra
 - be sure to backup your local .extra if it's been customized
after initial install the script updates git so it's ignored
but on rare occasions when this file is updated again
when you pull, you will get those changes. TODO a bun script
that automatically backups up .extra, before a pull, so you can deal with rare conflicts. Test and upate