feat: v0.4.26-gostty-kanata.0 ...
This commit starts an adoc about kanata config that is a wip. Externally
we added a plist so the karabiner daemon required by kanata does not
have to be started manually. In the future, we'd like some similar
mechanism for kanata debug or release cargo builds. However, for now
changes were made to aliases and local .extra as mentioned in the new
.cfg_KANATA.adoc such that we only now need to start kanata in debug
mode preferably in a tmux session so we can monitor it's state as it's a
pretty active project.

 - ALSO we started config for terminal ghostty where we expect changes
overtime. right now we are using a paid font so well experiment with
fall back when not on liscensed machines