fix: v0.4.100 RC.5 semver and code...
- avoiding hangs by 'lying to bash saying editor is :'

The Solution: The "Inline Override"

We have to force Git to behave non-interactively just for this one line.
We do that by setting two variables on the same line as the command.
This is the cleanest, most standard way to automate Git in scripts.

    GIT_MERGE_AUTOEDIT=no: Tells Git "Use the default 'Merge branch...'
message, don't ask me."

    GIT_EDITOR=":": The "Nuclear Option." If Git ignores the first
setting and tries to open an editor anyway, it tries to run :, which
instantly exits successfully. This makes a hang physically impossible.